Backend/requirements.txt
django>=5.0,<6.0
djangorestframework
pymysql
django-cors-headers
reportlab
